pyproject.toml
[project]
name = "prime-agent"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    # FastAPI and server
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.6",

    # Claude Agent SDK (includes anthropic as dependency)
    "claude-agent-sdk>=0.1.18",
    "anthropic>=0.75.0",

    # Git operations
    "gitpython>=3.1.46",

    # Utilities
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "types-PyYAML>=6.0.0",

    # Apple Push Notifications
    "aioapns>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "ruff>=0.14.10",
    "mypy>=1.19.1",
    "black>=25.12.0",
    "isort>=7.0.0",
    "pre-commit>=3.8.0",
    "pip-tools>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.pyc",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "G",      # flake8-logging-format
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "PERF",   # perflint
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
    "G004",   # logging format (f-strings are fine in Python 3)
    "PLW0603", # global statement (needed for FastAPI dependency injection pattern)
    "B008",   # function call in argument defaults (FastAPI Depends pattern)
    "ARG001", # unused function arguments (needed for FastAPI/interface signatures)
    "PLR0912", # too many branches
    "PLR0915", # too many statements
    "PTH123", # open() vs Path.open() (both are acceptable)
    "RUF012", # mutable class attributes (acceptable for constants)
    "RUF006", # asyncio.create_task (fire-and-forget pattern)
    "B904",   # raise from (not always needed)
    "DTZ",    # datetime timezone (tests use naive dates intentionally)
    "SIM105", # contextlib.suppress (explicit try-except-pass is clearer)
    "PLC0415", # import not at top-level (needed for fixtures in tests)
    "UP045",  # Optional vs X | None (Optional is more readable)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",    # unused function arguments
    "S101",   # use of assert
    "PLR2004", # magic values in tests
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "git.*",
    "claude_agent_sdk.*",
]
ignore_missing_imports = true
